requirements.txt
attrs==23.2.0
coverage==7.4.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==7.4.4
pytest-cov==4.1.0
python-dotenv==1.0.1
tomli==2.0.1
